+++ src/CosmoSim/__init__.py
@@ -11,6 +11,7 @@ import threading as th
 import os, sys
 
 import traceback
+from .Helper.Parameters import Parameters 
 
 __version__ = "2.8.0"